// src/enums/booking.ts
export enum BookingStatus {
  PENDING = 1,
  CANCELLED = 2,
  INTRANSIT = 3,
  COMPLETED = 4,
  ONBOARDED = 5
}

export enum BookingDirection {
  OFFICE = 1,
  HOME = 2
}

export enum BookingImportDirection {
  OFFICE = 1,
  HOME = 2
}

export enum BookingArrivalStatus {
  NOT_YET_ARRIVED = 1,
  ALREADY_ARRIVED = 2
}